+++ src/primitives/Breadcrumbs/BreadcrumbsItem.tsx
@@ -1,38 +1,26 @@
 import React from "react";
 
-import TypographyLink, { TypographyLinkProps } from "../Typography/TypographyLink";
+import { TypographyLinkProps } from "../Typography/TypographyLink";
 import Wrapper from "../Wrapper";
 import Typography from "../Typography";
 import Icon from "../Icon";
+import BreadcrumbsItemWrapper from "./BreadcrumbsItemWrapper";
 import {
-  ai,
-  backgroundColor,
-  borderRadius,
   bottom,
-  boxShadow,
   child,
   color,
   content,
-  disableDecoration,
-  disableOutline,
   display,
   flex,
-  focus,
-  fontSize,
-  fontWeight,
-  hover,
   left,
   lineHeight,
   marginBottom,
   marginRight,
-  padding,
   position,
   right,
   top,
-  transition,
 } from "../../styles";
-import { duration160 } from "../../constants/durations";
-import { Colors } from "constants/colors";
+import { Colors } from "../../constants/colors";
 
 export interface BreadcrumbsItemProps extends Omit<TypographyLinkProps, "children"> {
   text: string;
@@ -41,48 +29,10 @@ export interface BreadcrumbsItemProps extends Omit<TypographyLinkProps, "childre
   badgeColor: Colors;
 }
 
-interface BreadcrumbsWrapperProps {
-  isLastItem: boolean;
-  children?: React.ReactNode;
-}
-
 function BreadcrumbsItem({ to, text, isLastItem, withBadge, badgeColor }: BreadcrumbsItemProps) {
-  const breadcrumbsItemCommonStyles = [
-    position("relative"),
-    display("block"),
-    ai("center"),
-    padding(withBadge ? "2px 8px 3px 24px" : "2px 8px 3px"),
-    borderRadius(4),
-    fontSize(14),
-    lineHeight("1.5"),
-    fontWeight(500),
-  ];
-
-  const BreadcrumbsWrapper = ({ isLastItem, children }: BreadcrumbsWrapperProps) => {
-    if (isLastItem) {
-      return <Typography styles={breadcrumbsItemCommonStyles}>{children}</Typography>;
-    }
-
-    return (
-      <TypographyLink
-        to={to}
-        styles={[
-          breadcrumbsItemCommonStyles,
-          hover(backgroundColor("gray-blue/01")),
-          focus(boxShadow([0, 0, 0, 2, "blue/05"])),
-          transition(`background-color, box-shadow ${duration160}`),
-          disableOutline,
-          disableDecoration,
-        ]}
-      >
-        {children}
-      </TypographyLink>
-    );
-  };
-
   return (
     <Wrapper styles={[position("relative"), flex, lineHeight(1), marginRight(10), marginBottom(10)]}>
-      <BreadcrumbsWrapper isLastItem={isLastItem}>
+      <BreadcrumbsItemWrapper isLastItem={isLastItem} withBadge={withBadge} to={to}>
         {withBadge && (
           <Wrapper styles={[marginRight(8)]}>
             <Icon
@@ -101,12 +51,12 @@ function BreadcrumbsItem({ to, text, isLastItem, withBadge, badgeColor }: Breadc
             display("inline"),
             color("gray-blue/05"),
             !isLastItem &&
-              child([content("/"), position("absolute"), bottom(-1), right(-16), display("block")], "&:after"),
+              child([content("/"), position("absolute"), bottom(-2), right(-16), display("block")], "&:after"),
           ]}
         >
           {text}
         </Typography>
-      </BreadcrumbsWrapper>
+      </BreadcrumbsItemWrapper>
     </Wrapper>
   );
 }
