+++ src/primitives/Breadcrumbs/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
 import Wrapper from "../Wrapper";
-import { display, flexWrap, fullWidth } from "styles";
-import { Colors } from "constants/colors";
+import { flex, flexWrap, fullWidth } from "../../styles";
+import { Colors } from "../../constants/colors";
 
 export interface BreadcrumbsProps {
   children: React.ReactNode;
@@ -11,7 +11,7 @@ export interface BreadcrumbsProps {
 }
 
 function Breadcrumbs({ withBadge = false, badgeColor = "blue/05", children }: BreadcrumbsProps) {
-  return <Wrapper styles={[display("flex"), flexWrap, fullWidth]}>{children}</Wrapper>;
+  return <Wrapper styles={[flex, flexWrap, fullWidth]}>{children}</Wrapper>;
 }
 
 export default React.memo(Breadcrumbs);
